fix: use executable in PATH to run action
Github actions overrwides the workspace, and thus the action would
not work by simply calling bundler. In order to prevent issues arising
from this behavior, a helper script has been added, that switches
the current worskpace to /usr/src/app as it was intended to be.
On top it then executes bundler to run the dependabot script in the
correct context